# fix(android): improve performance of instant to ISO string