--- a/util/src/commonTest/kotlin/com/wire/kalium/util/DateTimeUtilTest.kt
+++ b/util/src/commonTest/kotlin/com/wire/kalium/util/DateTimeUtilTest.kt
@@ -1,9 +1,9 @@
-package com.wire.kalium.logic.util
+package com.wire.kalium.util
 
-import com.wire.kalium.util.DateTimeUtil
 import kotlinx.datetime.Instant
-import org.junit.Test
+import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
 class DateTimeUtilTest {
 
@@ -14,43 +14,43 @@ class DateTimeUtilTest {
 
     @Test
     fun givenAValidIsoDateTimeString_whenMatchingRegex_thenShouldSucceed() {
-        assert(regex.matches(isoDateTimeStringWith0Millis))
+        assertTrue(regex.matches(isoDateTimeStringWith0Millis))
     }
 
     @Test
     fun givenAnInvalidIsoDateTimeString_whenMatchingRegex_thenShouldSucceed() {
-        assert(!regex.matches("2022-12-20T17:30:00Z"))
-        assert(!regex.matches("2022-12-20T17:30:00.AZ"))
-        assert(!regex.matches("2022/12/20T17:30:00.000Z"))
-        assert(!regex.matches("2022-12-2017:30:00.000Z"))
-        assert(!regex.matches("2022-12-20T7:30:00Z"))
+        assertTrue(!regex.matches("2022-12-20T17:30:00Z"))
+        assertTrue(!regex.matches("2022-12-20T17:30:00.AZ"))
+        assertTrue(!regex.matches("2022/12/20T17:30:00.000Z"))
+        assertTrue(!regex.matches("2022-12-2017:30:00.000Z"))
+        assertTrue(!regex.matches("2022-12-20T7:30:00Z"))
     }
 
     @Test
     fun givenAnInstantWithMillisEqualTo0_whenParsingToIsoString_thenMillisShouldNotBeIgnored() {
         val instant = Instant.fromEpochMilliseconds(epochMillis)
         val result = DateTimeUtil.fromInstantToIsoDateTimeString(instant)
-        assert(regex.matches(result))
+        assertTrue(regex.matches(result))
     }
 
     @Test
     fun givenAnIsoDateTimeStringMillisEqualTo0_whenParsingToInstantAndBackToIsoString_thenMillisShouldNotBeIgnored() {
         val instant = Instant.parse(isoDateTimeStringWith0Millis)
         val result = DateTimeUtil.fromInstantToIsoDateTimeString(instant)
-        assert(regex.matches(result))
+        assertTrue(regex.matches(result))
     }
 
     @Test
     fun givenAnIsoDateTimeStringMillisEqualTo0_whenParsingToEpochMillisAndBackToIsoString_thenMillisShouldNotBeIgnored() {
         val epochMillis = DateTimeUtil.fromIsoDateTimeStringToEpochMillis(isoDateTimeStringWith0Millis)
         val result = DateTimeUtil.fromEpochMillisToIsoDateTimeString(epochMillis)
-        assert(regex.matches(result))
+        assertTrue(regex.matches(result))
     }
 
     @Test
     fun givenAnIsoDateTimeStringMillisEqualTo0_whenSubtractingMillis_thenMillisShouldNotBeIgnored() {
         val result = DateTimeUtil.minusMilliseconds(isoDateTimeStringWith0Millis, 1000)
-        assert(regex.matches(result))
+        assertTrue(regex.matches(result))
     }
 
     @Test
@@ -63,7 +63,7 @@ class DateTimeUtilTest {
     @Test
     fun whenGettingCurrentIsoDateTimeString_thenMillisShouldNotBeIgnored() {
         val result = DateTimeUtil.currentIsoDateTimeString()
-        assert(regex.matches(result))
+        assertTrue(regex.matches(result))
     }
 
     @Test