----
#### PR Submission Checklist for internal contributors

- The **PR Title**
  - [X] conforms to the style of semantic commits messages¹ supported in Wire's Github Workflow²
  - [X] answers the question: _If merged, this PR will: ..._ ³

- The **PR Description**
  - [X] is free of optional paragraphs and you have filled the relevant parts to the best of your ability
----

# What's new in this PR?

### Issues

Converting Instant to ISO 8601 string is not performant on Android at the moment.

### Causes

Whilst I can't exactly pinpoint to the exact cause, I didn't observe the same under-performance on JVM. Maybe the way Android implemented this API is different from regular JVMs.

### Solutions

Replace the date/time format with a default `DateFormatterBuilder` with a `appendInstant(int fractionalDigits)`, passing `3` to make sure it always adds 3 digits to the milliseconds.

See the benchmark below.

#### Also

Moved `DateUtilsTest` stuff to `utils` package.

### Testing

#### Test Coverage

- [X] I have added automated test to this contribution

I've added a benchmark for the future.

Running this benchmark, that transforms _one million_ random instants into ISO date time string:

##### Current implementation, using DateTimeFormatter
- 9.215483600s
- 8.698548s
- 8.898969300s
- 8.793911400s
- 9.084778500s

##### Doing Instant.toString()
- 855.682ms
- 904.086400ms
- 1.109003500s
- 1.080615400s
- 849.341700ms

##### Using DateTimeFormatter.appendInstant(3) to keep milliseconds
- 1.275490s
- 1.054792100s
- 1.021491900s
- 984.369100ms
- 1.196082400s

----
#### PR Post Merge Checklist for internal contributors

 - [X] If any soft of configuration variable was introduced by this PR, it has been added to the relevant documents and the CI jobs have been updated.
----
##### References
1. https://sparkbox.com/foundry/semantic_commit_messages
1. https://github.com/wireapp/.github#usage
1. E.g. `feat(conversation-list): Sort conversations by most emojis in the title #SQPIT-764`.